Make send_fd repeatable and reliable
Repeated calls to bpf_recv_fd would fail due the bpf_send_fd code only
sending one message. Turn this into a loop.

Related to the above loop, the SourceSocket destructor would hang
waiting for the sending thread to finish, but it would be blocked in
accept(). Fix that by calling shutdown() on the sock fd.

Signed-off-by: Brenden Blanco <bblanco@plumgrid.com>